pyproject.toml
[project]
name = "udaan-ai"
version = "1.0.0"
description = "agentic platform for personalized career & education guidance"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pre-commit>=4.0.0",
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.10.0",
    "google-generativeai>=0.8.5",
    "pydantic>=2.11.9",
]
